Web_App_Python/requirements.txt
asgiref==3.7.2
Django==4.2.7
psycopg2==2.9.9
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
